with_bugs

+++ search_engine.py
@@ -3,7 +3,7 @@ from .query_parser import Query
 from .corpus_manager import Document
 
 class SearchEngine:
-    def __call__(index: InvertedIndex, query: Query) -> dict[Document: int]:
+    def __call__(index: InvertedIndex, query: Query) -> dict[Document: float]:
         # ...
         pass
